+++ tests/test_cli.py
@@ -5,14 +5,16 @@ import subprocess
 import sys
 
 logger = logging.getLogger()
-package_dir = 'lib_travis'
-cli_filename = 'lib_travis_cli.py'
+package_dir = "lib_travis"
+cli_filename = "lib_travis_cli.py"
 
-path_cli_command = pathlib.Path(__file__).resolve().parent.parent / package_dir / cli_filename
+path_cli_command = (
+    pathlib.Path(__file__).resolve().parent.parent / package_dir / cli_filename
+)
 
 
-def call_cli_command(commandline_args: str = '') -> bool:
-    command = ' '.join([sys.executable, str(path_cli_command), commandline_args])
+def call_cli_command(commandline_args: str = "") -> bool:
+    command = " ".join([sys.executable, str(path_cli_command), commandline_args])
     try:
         subprocess.run(command, shell=True, check=True)
     except subprocess.CalledProcessError:
@@ -23,17 +25,19 @@ def call_cli_command(commandline_args: str = '') -> bool:
 def test_cli_commands() -> None:
     # due to a bug in python 3.8.1 with setup.py test on travis we need to cancel the click tests there !
     if sys.version_info < (3, 8, 1) or sys.version_info >= (3, 8, 2):
-        assert not call_cli_command('--unknown_option')
-        assert call_cli_command('--version')
-        assert call_cli_command('-h')
-        assert call_cli_command('info')
-        assert call_cli_command('--traceback info')
-        assert call_cli_command('get_branch')
+        assert not call_cli_command("--unknown_option")
+        assert call_cli_command("--version")
+        assert call_cli_command("-h")
+        assert call_cli_command("info")
+        assert call_cli_command("--traceback info")
+        assert call_cli_command("get_branch")
 
         assert call_cli_command('run test "echo test"')
-        assert not call_cli_command('run description "unknown command" --retry=3 --sleep=0')
+        assert not call_cli_command(
+            'run description "unknown command" --retry=3 --sleep=0'
+        )
 
-        assert call_cli_command('install --dry-run')
-        assert call_cli_command('script --dry-run')
-        assert call_cli_command('after_success --dry-run')
-        assert call_cli_command('deploy --dry-run')
+        assert call_cli_command("install --dry-run")
+        assert call_cli_command("script --dry-run")
+        assert call_cli_command("after_success --dry-run")
+        assert call_cli_command("deploy --dry-run")